+++ bin/extract-env-vars.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../lib/cli').execute(process.argv);
